requirements.txt
Flask==1.1.2
PyMySQL==0.10.1
PyYAML==5.4.1
